requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
requests>=2.31.0
httpx[http2]>=0.26.0 